requirements.txt
pandas==2.0.3
ta==0.10.2
psycopg2-binary==2.9.9
